requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
pydantic==2.11.1
python-dotenv==1.1.0
aiofiles>=23.0.0
google-genai>=1.0.0
mcp>=1.5.0
google-adk>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.22.0
pre-commit>=3.3.2
pytest-html>=3.2.0
deepeval>=2.0.0
pytest-timeout>=2.2.4
